[Grupo 16] feat: criando navbar de products

+++ back-end/src/database/migrations/20230328204738-products.js
@@ -3,7 +3,7 @@ module.exports = {
     await queryInterface.createTable('products', {
       id: { type: Sequelize.INTEGER, allowNull: false, primaryKey: true, autoIncrement: true },
       name: { type: Sequelize.STRING, allowNull: false },
-      price: { type: Sequelize.FLOAT, allowNull: false },
+      price: { type: Sequelize.DECIMAL(4, 2), allowNull: false },
       urlImage: { type: Sequelize.STRING, allowNull: false, field: 'url_image' },
     });
   },
